+++ examples/KWK_getcheckdata.py
@@ -2,7 +2,6 @@
 
 import os
 import pandas as pd
-import datetime as dt
 import matplotlib.pyplot as plt
 plt.close('all')
 import kenmerkendewaarden as kw
@@ -31,7 +30,6 @@ write_stations_table = False
 start_date = pd.Timestamp(1870, 1, 1, tz="UTC+01:00")
 end_date = pd.Timestamp(2024, 1, 1, tz="UTC+01:00")
 
-# dir_base = r'p:\11208031-010-kenmerkende-waarden-k\work'
 dir_base = r"p:\11210325-005-kenmerkende-waarden\work"
 dir_meas = os.path.join(dir_base, f"measurements_wl_{start_date.strftime('%Y%m%d')}_{end_date.strftime('%Y%m%d')}")
 os.makedirs(dir_meas, exist_ok=True)
@@ -68,21 +66,21 @@ for stat_remove in stations_skip:
 if retrieve_meas_amount:
     kw.retrieve_measurements_amount(dir_output=dir_meas_amount, station_list=station_list, 
                                     start_date=start_date, end_date=end_date,
-                                    extremes=False)
+                                    quantity="meas_wl")
     kw.retrieve_measurements_amount(dir_output=dir_meas_amount, station_list=station_list, 
                                     start_date=start_date, end_date=end_date,
-                                    extremes=True)
+                                    quantity="meas_ext")
 
 
 ### PLOT MEASUREMENTS AMOUNT
 if plot_meas_amount:
-    df_amount_ts = kw.read_measurements_amount(dir_output=dir_meas_amount, extremes=False)
-    df_amount_ext = kw.read_measurements_amount(dir_output=dir_meas_amount, extremes=True)
+    df_amount_wl = kw.read_measurements_amount(dir_output=dir_meas_amount, quantity="meas_wl")
+    df_amount_ext = kw.read_measurements_amount(dir_output=dir_meas_amount, quantity="meas_ext")
     
     file_plot = os.path.join(dir_meas_amount, "data_amount")
     
-    fig, ax = kw.plot_measurements_amount(df_amount_ts, relative=True)
-    fig.savefig(file_plot + "_ts_pcolormesh_relative", dpi=200)
+    fig, ax = kw.plot_measurements_amount(df_amount_wl, relative=True)
+    fig.savefig(file_plot + "_wl_pcolormesh_relative", dpi=200)
     fig, ax = kw.plot_measurements_amount(df_amount_ext, relative=True)
     fig.savefig(file_plot + "_ext_pcolormesh_relative", dpi=200)
     
@@ -93,18 +91,18 @@ for current_station in station_list:
     if not retrieve_meas:
         continue
     
-    kw.retrieve_measurements(dir_output=dir_meas, station=current_station, extremes=False,
+    kw.retrieve_measurements(dir_output=dir_meas, station=current_station, quantity="meas_wl",
                              start_date=start_date, end_date=end_date)
-    kw.retrieve_measurements(dir_output=dir_meas, station=current_station, extremes=True,
+    kw.retrieve_measurements(dir_output=dir_meas, station=current_station, quantity="meas_ext",
                              start_date=start_date, end_date=end_date)
 
 
 
 ### CREATE SUMMARY
 if derive_stats:
-    stats_ts = kw.derive_statistics(dir_output=dir_meas, station_list=station_list, extremes=False)
-    stats_ext = kw.derive_statistics(dir_output=dir_meas, station_list=station_list, extremes=True)
-    stats_ts.to_csv(os.path.join(dir_meas,'data_summary_ts.csv'))
+    stats_wl = kw.derive_statistics(dir_output=dir_meas, station_list=station_list, quantity="meas_wl")
+    stats_ext = kw.derive_statistics(dir_output=dir_meas, station_list=station_list, quantity="meas_ext")
+    stats_wl.to_csv(os.path.join(dir_meas,'data_summary_wl.csv'))
     stats_ext.to_csv(os.path.join(dir_meas,'data_summary_ext.csv'))
 
 
@@ -117,15 +115,15 @@ for current_station in station_list:
     print(f'plotting timeseries data for {current_station}')
     
     # load data
-    df_meas = kw.read_measurements(dir_output=dir_meas, station=current_station, extremes=False)
-    df_ext = kw.read_measurements(dir_output=dir_meas, station=current_station, extremes=True)
+    df_meas = kw.read_measurements(dir_output=dir_meas, station=current_station, quantity="meas_wl")
+    df_ext = kw.read_measurements(dir_output=dir_meas, station=current_station, quantity="meas_ext")
     
     # create and save figure
     fig,(ax1, ax2) = kw.plot_measurements(df_meas=df_meas, df_ext=df_ext)
     file_wl_png = os.path.join(dir_meas,f'ts_{current_station}.png')
     ax1.set_xlim(pd.Timestamp(start_date), pd.Timestamp(end_date)) # entire period
     fig.savefig(file_wl_png.replace('.png','_alldata.png'))
-    ax1.set_xlim(dt.datetime(2000,1,1),dt.datetime(2024,1,1)) # period of interest
+    ax1.set_xlim(pd.Timestamp(2000,1,1),pd.Timestamp(2024,1,1)) # period of interest
     fig.savefig(file_wl_png.replace('.png','_2000_2024.png'))
     plt.close(fig)
 
@@ -147,7 +145,7 @@ if plot_stations:
 if write_stations_table:
     # TODO: consider making retrieve_catalog public
     from kenmerkendewaarden.data_retrieve import retrieve_catalog
-    locs_meas_ts_all, _, _ = retrieve_catalog(crs=4326)
-    locs_ts = locs_meas_ts_all.loc[locs_meas_ts_all.index.isin(station_list)]
+    locs_meas_wl_all, _, _, _ = retrieve_catalog(crs=4326)
+    locs_wl = locs_meas_wl_all.loc[locs_meas_wl_all.index.isin(station_list)]
     file_csv = os.path.join(dir_base, "station_locations.csv")
-    locs_ts[["Locatie_MessageID","X","Y","Coordinatenstelsel","Naam"]].to_csv(file_csv)
+    locs_wl[["Locatie_MessageID","X","Y","Coordinatenstelsel","Naam"]].to_csv(file_csv)